# systems/crawler/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AACV System API
  version: 7.1.0
  description: 本契约描述认证、用户、审计、双源采集、MySQL 数据治理、规范目录、事务 Outbox、Neo4j 可重建投影、受限图查询、MySQL 聚合统计、异步导出和管理员运维接口；不包含服务器部署能力。
servers:
  - url: http://127.0.0.1:8080
    description: Windows 本地开发后端
...
paths:
...
  /api/v1/sources:
    get:
      tags: [Sources]
      summary: 分页查询受控数据源
      operationId: listSources
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 数据源分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSourcePage'
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Sources]
      summary: 创建唯一的受控 OpenAlex 或 Crossref 数据源配置
      operationId: createSource
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SourceConfigurationRequest'}
      responses:
        '201':
          description: 数据源已创建
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataSource'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/sources/{sourceId}:
    get:
      tags: [Sources]
      summary: 查询数据源详情
      operationId: getSource
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/SourceId'
      responses:
        '200':
          description: 数据源详情
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataSource'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Sources]
      summary: 更新数据源受控配置
      operationId: updateSource
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/SourceId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SourceConfigurationRequest'}
      responses:
        '200':
          description: 数据源已更新
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataSource'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/sources/{sourceId}/enable:
    post:
      tags: [Sources]
      summary: 启用数据源
      operationId: enableSource
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/SourceId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody: {$ref: '#/components/requestBodies/VersionRequest'}
      responses:
        '200':
          description: 数据源已启用
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataSource'}
  /api/v1/sources/{sourceId}/disable:
    post:
      tags: [Sources]
      summary: 停用数据源
      operationId: disableSource
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/SourceId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody: {$ref: '#/components/requestBodies/VersionRequest'}
      responses:
        '200':
          description: 数据源已停用
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataSource'}
  /api/v1/sources/{sourceId}/probe:
    post:
      tags: [Sources]
      summary: 对来源对应的固定官方地址执行无数据写入探测
      operationId: probeSource
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/SourceId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: 探测结果
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SourceProbe'}
...
  /api/v1/graph/types:
    get:
      tags: [Graph]
      summary: 读取 MySQL 图谱类型配置（GRAPH_READ）
      operationId: listGraphTypes
      security: [{sessionCookie: []}]
      responses:
        '200': {description: 类型配置列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GraphTypeDefinition'}}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/graph/types/{kind}/{code}:
    put:
      tags: [Graph]
      summary: 保存类型名称、颜色、尺寸与类型审核状态（GRAPH_SYNC_MANAGE）
      description: 要求 CSRF 与当前版本，成功后版本加1并记录审计。只更新受支持的已有类型；类型审核不代表具体实体或关系已审核。
      operationId: updateGraphType
      security: [{sessionCookie: []}]
      parameters:
        - {name: kind, in: path, required: true, schema: {type: string, enum: [NODE, RELATIONSHIP]}}
        - {name: code, in: path, required: true, schema: {type: string, pattern: '^[A-Z_]{1,32}$'}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GraphTypeUpdate'}
      responses:
        '200': {description: 已保存的配置, content: {application/json: {schema: {$ref: '#/components/schemas/GraphTypeDefinition'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {description: 类型配置版本已变化，请刷新后重试}
  /api/v1/graph/overview:
    get:
      tags: [Graph]
      summary: 读取受限作者作品概览
      operationId: getGraphOverview
      description: 要求 GRAPH_READ。读取 Neo4j 受管作者、作品和 AUTHORED，优先返回创作关系，再补孤立节点；查询超时3秒，候选路径最多 nodeLimit*4+1。MySQL 提供 typeDefinitions，合作依据仅为本次返回作品，最多1000条派生边。无数据返回空数组，rootNodeId 为空字符串；统计仅针对当前结果。
      security: [{sessionCookie: []}]
      parameters:
        - {name: nodeLimit, in: query, schema: {type: integer, minimum: 1, maximum: 300, default: 300}}
      responses:
        '200': {description: 有界网络或空网络, content: {application/json: {schema: {$ref: '#/components/schemas/GraphResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/GraphUnavailable'}
        '504': {$ref: '#/components/responses/GraphQueryTimeout'}
  /api/v1/graph/subgraph:
    get:
      tags: [Graph]
      summary: 加载受限局部子图
      operationId: getGraphSubgraph
      description: 拓扑来自 Neo4j，typeDefinitions 来自 MySQL。派生合作关系仅以当前返回子图中的共同作品为依据，最多1000条，达到上限时给出截断提示。
      security: [{sessionCookie: []}]
      parameters:
        - {name: centerType, in: query, required: true, schema: {$ref: '#/components/schemas/GraphNodeType'}}
        - {name: includeCoauthors, in: query, schema: {type: boolean, default: false}}
        - {name: centerId, in: query, required: true, schema: {type: integer, format: int64, minimum: 1}}
        - {name: depth, in: query, schema: {type: integer, minimum: 1, maximum: 2, default: 1}}
        - {name: nodeLimit, in: query, schema: {type: integer, minimum: 1, maximum: 300, default: 100}}
        - {name: relationshipTypes, in: query, schema: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/GraphRelationshipType'}}}
        - {name: nodeTypes, in: query, schema: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/GraphNodeType'}}}
        - {name: publicationYearFrom, in: query, schema: {type: integer, minimum: 1000, maximum: 9999}}
        - {name: publicationYearTo, in: query, schema: {type: integer, minimum: 1000, maximum: 9999}}
        - {name: achievementTypes, in: query, schema: {type: array, uniqueItems: true, items: {type: string, maxLength: 64}}}
      responses:
        '200': {description: 局部子图, content: {application/json: {schema: {$ref: '#/components/schemas/GraphResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/GraphUnavailable'}
        '504': {$ref: '#/components/responses/GraphQueryTimeout'}
  /api/v1/graph/path:
    get:
      tags: [Graph]
      summary: 查询一条确定性的受限最短路径
      operationId: getGraphPath
      security: [{sessionCookie: []}]
      parameters:
        - {name: sourceType, in: query, required: true, schema: {$ref: '#/components/schemas/GraphNodeType'}}
        - {name: sourceId, in: query, required: true, schema: {type: integer, format: int64, minimum: 1}}
        - {name: targetType, in: query, required: true, schema: {$ref: '#/components/schemas/GraphNodeType'}}
        - {name: targetId, in: query, required: true, schema: {type: integer, format: int64, minimum: 1}}
        - {name: maxHops, in: query, schema: {type: integer, minimum: 1, maximum: 6, default: 6}}
      responses:
        '200': {description: 最短路径, content: {application/json: {schema: {$ref: '#/components/schemas/GraphResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/GraphUnavailable'}
        '504': {$ref: '#/components/responses/GraphQueryTimeout'}
  /api/v1/graph/sync-status:
    get:
      tags: [Graph Operations]
      summary: 查询图投影同步状态
      operationId: getGraphSyncStatus
      security: [{sessionCookie: []}]
      responses:
        '200': {description: 同步状态, content: {application/json: {schema: {$ref: '#/components/schemas/GraphSyncStatus'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/operations/graph-events:
    get:
      tags: [Graph Operations]
      summary: 分页查询安全图同步事件
      operationId: getGraphEvents
      security: [{sessionCookie: []}]
      parameters:
        - {name: status, in: query, schema: {$ref: '#/components/schemas/GraphOutboxStatus'}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200': {description: 事件分页, content: {application/json: {schema: {$ref: '#/components/schemas/GraphEventPage'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/operations/graph-events/{eventId}/replay:
    post:
      tags: [Graph Operations]
      summary: 按MySQL当前状态重放单个失败事件
      operationId: replayGraphEvent
      security: [{sessionCookie: []}]
      parameters:
        - {name: eventId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/CsrfHeader'}
      responses:
        '202': {description: 已创建新的刷新事件, content: {application/json: {schema: {$ref: '#/components/schemas/GraphEvent'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
    SourceId:
      name: sourceId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  requestBodies:
    VersionRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VersionRequest'
  responses:
...
    BadRequest:
      description: 请求校验失败
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 未登录、凭据错误或会话失效
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: 权限不足或 CSRF Token 无效
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 资源状态、并发限制、用户名或乐观锁版本冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 请求的资源不存在
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    GraphUnavailable:
      description: Neo4j不可用或图投影正在重建
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    GraphQueryTimeout:
      description: Neo4j图查询超过服务端限定时间
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    VersionRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version:
          type: integer
          format: int64
          minimum: 0
...
    SourceConfigurationRequest:
      type: object
      additionalProperties: false
      required: [requestsPerSecond, maxConcurrency, connectTimeoutSeconds, responseTimeoutSeconds, maxRetries, maxResponseBytes, complianceNote]
      properties:
        sourceType: {type: [string, 'null'], enum: [OPENALEX, CROSSREF, null], description: 创建时省略则兼容为 OPENALEX；更新时保持既有来源类型}
        requestsPerSecond: {type: integer, minimum: 1, maximum: 10}
        maxConcurrency: {type: integer, minimum: 1, maximum: 4}
        connectTimeoutSeconds: {type: integer, minimum: 1, maximum: 30}
        responseTimeoutSeconds: {type: integer, minimum: 1, maximum: 120}
        maxRetries: {type: integer, minimum: 0, maximum: 5}
        maxResponseBytes: {type: integer, minimum: 1024, maximum: 20971520}
        complianceNote: {type: string, minLength: 1, maxLength: 1000}
        version: {type: integer, format: int64, minimum: 0}
    DataSource:
      type: object
      required: [id, sourceCode, sourceType, baseUri, enabled, requestsPerSecond, maxConcurrency, connectTimeoutSeconds, responseTimeoutSeconds, maxRetries, maxResponseBytes, complianceNote, consecutiveFailures, version]
      properties:
        id: {type: integer, format: int64}
        sourceCode: {type: string, enum: [OPENALEX, CROSSREF]}
        sourceType: {type: string, enum: [OPENALEX, CROSSREF]}
        baseUri: {type: string, format: uri, enum: ['https://api.openalex.org', 'https://api.crossref.org']}
        enabled: {type: boolean}
        requestsPerSecond: {type: integer}
        maxConcurrency: {type: integer}
        connectTimeoutSeconds: {type: integer}
        responseTimeoutSeconds: {type: integer}
        maxRetries: {type: integer}
        maxResponseBytes: {type: integer}
        complianceNote: {type: string}
        lastSuccessAt: {type: [string, 'null'], format: date-time}
        lastFailureAt: {type: [string, 'null'], format: date-time}
        consecutiveFailures: {type: integer}
        version: {type: integer, format: int64}
    DataSourcePage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/DataSource'}}
    SourceProbe:
      type: object
      required: [reachable, rateLimitSummary, checkedAt]
      properties:
        reachable: {type: boolean}
        statusCode: {type: [integer, 'null']}
        errorCategory: {type: [string, 'null']}
        rateLimitSummary: {type: object, additionalProperties: {type: string}}
        checkedAt: {type: string, format: date-time}
...
    PageMetadata:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
    GraphNodeType:
      type: string
      enum: [ACHIEVEMENT, AUTHOR, INSTITUTION, VENUE, TOPIC]
    GraphRelationshipType:
      type: string
      enum: [AUTHORED, AFFILIATED_WITH, PUBLISHED_IN, HAS_TOPIC, CITES]
    GraphOutboxStatus:
      type: string
      enum: [PENDING, PROCESSING, SUCCEEDED, DEAD]
    GraphNode:
      type: object
      required: [id, businessId, type, label, properties]
      properties:
        id: {type: string, description: 由固定节点类型和businessId组成的稳定响应ID}
        businessId: {type: string}
        type: {$ref: '#/components/schemas/GraphNodeType'}
        label: {type: string}
        properties: {type: object, additionalProperties: true, description: 仅包含展示所需的白名单摘要属性}
    GraphEdge:
      type: object
      required: [id, type, source, target, properties]
      properties:
        id: {type: string}
        type: {type: string, enum: [AUTHORED, AFFILIATED_WITH, PUBLISHED_IN, HAS_TOPIC, CITES, COAUTHORED]}
        source: {type: string}
        target: {type: string}
        properties: {type: object, additionalProperties: true}
    GraphAppliedLimits:
      type: object
      required: [depth, nodeLimit, maxHops]
      properties:
        depth: {type: integer, minimum: 0, maximum: 2}
        nodeLimit: {type: integer, minimum: 1, maximum: 300}
        maxHops: {type: integer, minimum: 0, maximum: 6}
    GraphResponse:
      type: object
      required: [nodes, edges, rootNodeId, truncated, narrowingSuggestion, appliedLimits, syncedAt, projectionLagSeconds, traceId]
      properties:
        typeDefinitions: {type: array, items: {$ref: '#/components/schemas/GraphTypeDefinition'}}
        nodes: {type: array, maxItems: 300, items: {$ref: '#/components/schemas/GraphNode'}}
        edges: {type: array, items: {$ref: '#/components/schemas/GraphEdge'}}
        rootNodeId: {type: string}
        truncated: {type: boolean}
        narrowingSuggestion: {type: [string, 'null']}
        appliedLimits: {$ref: '#/components/schemas/GraphAppliedLimits'}
        syncedAt: {type: [string, 'null'], format: date-time}
        projectionLagSeconds: {type: [integer, 'null'], format: int64, minimum: 0}
        traceId: {type: string}
    GraphTypeUpdate:
      type: object
      required: [displayName, color, size, reviewStatus, version]
      properties:
        displayName: {type: string, minLength: 1, maxLength: 64}
        color: {type: string, pattern: '^#[0-9a-fA-F]{6}$'}
        size: {type: integer, minimum: 1, maximum: 96, description: NODE 为16–96px直径，RELATIONSHIP 为1–8px线宽}
        reviewStatus: {type: string, enum: [PENDING, APPROVED, REJECTED]}
        version: {type: integer, format: int64, minimum: 0}
    GraphTypeDefinition:
      allOf:
        - {$ref: '#/components/schemas/GraphTypeUpdate'}
        - type: object
          required: [kind, code]
          properties:
            kind: {type: string, enum: [NODE, RELATIONSHIP]}
            code: {type: string, enum: [ACHIEVEMENT, AUTHOR, INSTITUTION, VENUE, TOPIC, AUTHORED, AFFILIATED_WITH, PUBLISHED_IN, HAS_TOPIC, CITES, COAUTHORED]}
    GraphSyncStatus:
      type: object
      required: [neo4jAvailable, schemaVersion, pendingCount, processingCount, deadCount, oldestPendingAgeSeconds, lastSucceededAt, lagThresholdExceeded, rebuildInProgress]
      properties:
        neo4jAvailable: {type: boolean}
        schemaVersion: {type: [integer, 'null']}
        pendingCount: {type: integer, format: int64, minimum: 0}
        processingCount: {type: integer, format: int64, minimum: 0}
        deadCount: {type: integer, format: int64, minimum: 0}
        oldestPendingAgeSeconds: {type: [integer, 'null'], format: int64, minimum: 0}
        lastSucceededAt: {type: [string, 'null'], format: date-time}
        lagThresholdExceeded: {type: boolean}
        rebuildInProgress: {type: boolean}
    GraphEvent:
      type: object
      required: [eventId, achievementId, desiredVersion, eventType, status, attempts, nextAttemptAt, createdAt, updatedAt]
      properties:
        eventId: {type: string, format: uuid}
        achievementId: {type: integer, format: int64}
        desiredVersion: {type: integer, format: int64, minimum: 1}
        eventType: {type: string, enum: [REFRESH]}
        status: {$ref: '#/components/schemas/GraphOutboxStatus'}
        attempts: {type: integer, minimum: 0}
        nextAttemptAt: {type: string, format: date-time}
        errorCode: {type: [string, 'null']}
        errorSummary: {type: [string, 'null']}
        replayOfEventId: {type: [string, 'null'], format: uuid}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        completedAt: {type: [string, 'null'], format: date-time}
    GraphEventPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/GraphEvent'}}
...
    Problem:
      type: object
      required: [type, title, status, detail, instance, errorCode, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errorCode:
          type: string
        traceId:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...